feat(process): Impl kill subcommand
Kill subcommand is implemented by adding a new module `process` to the project.

This module contains the actual logic that is used by the `hpm` binary
crate.

The actual "kill" functionality is implemented by calling `systemctl
poweroff`.

The entrypoint of the lib crate is updated to call `process::kill()`
child process.

The child process stdout stream is propagated to the inherited
`std::io::stdout()` stream.

The child process stderr stream is propagated to the binary crate.

Potential errors from the entrypoint:

- `process::Error` signifies the potential errors that may arise from
interacting with child process.
- `process:Process` is the main struct that controls how the child processes are interacted by the clients.

Dependencies:

The crate `which` is added to the project to validate the child
processes.

Signed-off-by: acikgozb <acikgozb@proton.me>