feat(interactive): Impl -i (interactive mode)
The mode is implemented by creating a prompt for the user. Upon
selecting the appropriate option (u8), `hpm` acts like it received an
argument and continues with the regular flow.

Interactive mode only affects the subcommand decision, there is no
difference between interactive and non-interactive mode in other parts
of `hpm`.

Signed-off-by: acikgozb <acikgozb@proton.me>